+++ src/app/factories/pets/useCases/pets.controller.ts
@@ -1,10 +1,12 @@
+import { IPetsController } from '@app/factories/repositories/pets.repository';
+import { AppError } from '@domain/errors/AppError';
+import { STATUS } from '@domain/helpers/constants';
 import { formatEmptyValues } from '@domain/helpers/utils/formatEmptyValues';
 import { IPetModel } from '@domain/models/pets';
 import { IUploadFile, uploadFile } from '@infra/upload';
 import { splitToDeleteS3Object } from '@infra/upload/fileUpload';
 import { Request, Response } from 'express';
 
-import { IPetsController } from './IPetsController';
 import PetsService from './pets.service';
 
 class PetsController implements IPetsController {
@@ -17,7 +19,7 @@ class PetsController implements IPetsController {
     const { Location } = await uploadFile(uploadedImage) as unknown as IUploadFile;
     const pet = await PetsService().create({ ...data, image: Location });
 
-    res.status(201).json({ pet });
+    res.status(STATUS.CREATED_CONTENT).json(pet);
   }
 
   async findAll(req: Request, res: Response) {
@@ -25,7 +27,7 @@ class PetsController implements IPetsController {
 
     const pets = await PetsService().findAll(String(orderBy));
 
-    res.status(200).json(pets);
+    res.status(STATUS.SUCCESS).json(pets);
   }
 
   async findById(req: Request, res: Response) {
@@ -33,17 +35,22 @@ class PetsController implements IPetsController {
 
     const pet = await PetsService().findById(id);
 
-    res.status(200).json(pet);
+    res.status(STATUS.SUCCESS).json(pet);
   }
 
   async delete(req: Request, res: Response) {
     const { id } = req.params;
 
     const pet = await PetsService().findById(id);
+
+    if (!pet) {
+      throw new AppError('Animal não encontrado', STATUS.NOT_FOUND);
+    }
+
     await PetsService().delete(id);
     splitToDeleteS3Object(pet.image);
 
-    res.status(204).json({ message: 'O animal foi deletado' });
+    res.status(STATUS.EMPTY_CONTENT).send();
   }
 }
